requirements.txt
transformers
transformers[sentencepiece]
datasets
ensure
sacrebleu
rouge_score
py7zr
pandas
nltk
tqdm
pyYAML
matplotlib
torch
notebook
boto3
mypy-boto3-s3
python-box==7.0.1
ensure==1.0.2
fastapi==0.78.0
uvicorn==0.18.3
Jinja2==3.1.2
-e .

